Made separate files
For easier navigations in project and for his potential
